# packages/api/src/generated/mdr.schemas.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * MDR Adviser API
 * API for the Medical Device Regulation (MDR) Adviser platform.
 * OpenAPI spec version: 1.0.0
 */
export type UserRole = typeof UserRole[keyof typeof UserRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserRole = {
  admin: 'admin',
  user: 'user',
  pro: 'pro',
} as const;

export type User = {
  firstName: string;
  lastName: string;
  role: UserRole;
};

export type SigninResponse = {
  user: User;
  accessToken: string;
  refreshToken: string;
};

export type SignInBody = {
  email: string;
  password: string;
};